refactor(reporting): split section_generators god-file into sections/ submodules
Decompose the 1146-line section_generators.py into 7 focused submodules under
reporting/sections/ (factpack, portfolio_summary, holdings, analysis, discovery,
sentiment, macro), each <300 lines. section_generators.py becomes a re-export
shim so all existing imports (incl. the private helpers tests reach for) keep
working. Behavior-preserving: 237 reporting tests pass unchanged.
