+++ src/finwiz/reporting/sections/__init__.py
@@ -1,0 +1,5 @@
+"""Report section generators, split by report area.
+
+Public ``generate_*`` functions (and a few helpers used by tests) are re-exported
+from ``finwiz.reporting.section_generators`` for backward compatibility.
+"""
