+++ src/finwiz/reporting/sections/holdings.py
@@ -1,0 +1,236 @@
+"""Holdings analysis table and strategic recommendations sections."""
+
+from __future__ import annotations
+
+from html import escape
+from typing import Any
+
+from finwiz.reporting.sections.factpack import _fact_pack_provenance_footer
+from finwiz.schemas.portfolio_review import HoldingDecision
+
+
+def _get_recommendation_badge(grade: str, recommended_action: str | None) -> str:
+    """Get recommendation badge HTML based on grade."""
+    if grade in ["A+", "A"]:
+        return '<span class="badge badge-buy">BUY</span>'
+    if grade in ["D", "F", "N/A"]:
+        return '<span class="badge badge-sell">SELL</span>'
+    if grade in ["B+", "B", "C+", "C"]:
+        return '<span class="badge badge-hold">HOLD</span>'
+
+    # Fallback to recommended_action
+    if recommended_action and "BUY" in recommended_action:
+        return '<span class="badge badge-buy">BUY</span>'
+    if recommended_action and "SELL" in recommended_action:
+        return '<span class="badge badge-sell">SELL</span>'
+    return '<span class="badge badge-hold">HOLD</span>'
+
+
+def _get_deep_analysis_link(ticker: str, asset_class: str) -> str:
+    """Generate relative link to deep analysis report if available.
+
+    Args:
+        ticker: The ticker symbol
+        asset_class: The asset class (stock, etf, crypto)
+
+    Returns:
+        Relative path to deep analysis HTML file
+
+    Note:
+        Reports are now stored in unified structure: output/{asset_class}/{ticker}_report.html
+        Generated on-the-fly by DeepAnalysisOrchestrator._store_enriched_analysis()
+    """
+    asset_class_lower = asset_class.lower()
+    # Unified output structure: output/stock/, output/etf/, output/crypto/
+    return f"{asset_class_lower}/{ticker}_report.html"
+
+
+def _format_target_cell(price_targets: Any, kind: str) -> str:
+    """Render one of the two ADR-011 compact columns ('target' or 'sell').
+
+    Returns ``<td class="muted">—</td>`` when there is no usable price_targets.
+    Otherwise produces ``<td>$X.XX<br><small>±Y.Y%</small></td>`` with HTML
+    escaping for safety. The ± delta is computed against ``price_targets.current_price``.
+    """
+    if price_targets is None:
+        return '<td class="muted">—</td>'
+    current = getattr(price_targets, "current_price", None)
+    if current is None or current <= 0:
+        return '<td class="muted">—</td>'
+    value = price_targets.buy_target_primary if kind == "target" else price_targets.sell_target_primary
+    if value is None:
+        return '<td class="muted">—</td>'
+    delta_pct = (value - current) / current * 100
+    sign = "+" if delta_pct >= 0 else "−"  # use Unicode minus for tasteful negatives
+    abs_pct = abs(delta_pct)
+    # f-string formatted floats — no operator-supplied content, but escape for defense in depth.
+    return f"<td>${value:,.2f}<br><small>{sign}{abs_pct:.1f}%</small></td>"
+
+
+def _confidence_badge(holding: HoldingDecision) -> str:
+    """Render the amber 'Insight IA indisponible' badge when confidence is low.
+
+    Returns empty string when confidence is high (no badge needed).
+    """
+    if getattr(holding, "confidence", "high") == "low":
+        return '<span class="badge-amber" title="Insight IA indisponible — proxy quantitatif Python affiché">⚠️ Insight IA indisponible</span>'
+    return ""
+
+
+def _render_holding_row(holding: HoldingDecision) -> str:
+    """Render a single holding as an HTML table row.
+
+    Pending holdings (grade N/A) get the 'Analyse en attente' placeholder.
+    Analyzed holdings with low confidence get an amber badge next to the grade.
+    """
+    grade = holding.grade or "N/A"
+    ticker = holding.ticker or "N/A"
+    name = holding.name or "Unknown"
+    asset_class = (holding.asset_class or "unknown").upper()
+    is_pending = grade == "N/A"
+
+    # Escape user-derived strings before HTML interpolation. `name` is free-form
+    # operator input; `ticker` is schema-validated to [A-Z0-9:.\-^=] but we
+    # escape anyway as defense in depth at the render boundary.
+    ticker_safe = escape(ticker)
+    name_safe = escape(name)
+
+    if is_pending:
+        # Don't link pending tickers to {ticker}_report.html — that file
+        # was never generated by _store_enriched_analysis (no analysis
+        # ran). Plain text avoids the broken-link 404.
+        ticker_plain_html = f"<strong>{ticker_safe}</strong>"
+        # Surface the *specific* upstream rationale (e.g. "Analyse interrompue :
+        # crew dépassé 900s — voir logs", "Analyse interrompue : circuit
+        # breaker ouvert ...") so users can tell apart timeout vs breaker vs
+        # missing-volatility vs generic-failure rows. Falls back to the
+        # original generic text only when no rationale was produced.
+        pending_msg = holding.rationale_bullets[0] if holding.rationale_bullets else "Analyse approfondie non disponible — relancer l'analyse pour obtenir un verdict."
+        pending_msg_safe = escape(pending_msg)
+        return f"""
+        <tr class="row-pending">
+          <td>{ticker_plain_html}<br><small>{name_safe}</small></td>
+          <td>{asset_class}</td>
+          <td class="grade-na" title="Deep analysis did not run for this holding"><strong>⏳ Analyse en attente</strong></td>
+          <td class="muted">—</td>
+          <td class="muted">—</td>
+          <td><small class="muted">{pending_msg_safe}</small></td>
+          <td class="muted">—</td>
+          <td class="muted">—</td>
+        </tr>"""
+
+    deep_analysis_link = _get_deep_analysis_link(ticker, holding.asset_class or "stock")
+    deep_analysis_link_safe = escape(deep_analysis_link, quote=True)
+    ticker_html = f'<a href="{deep_analysis_link_safe}" class="ticker-link" title="View detailed analysis for {ticker_safe}">{ticker_safe}</a>'
+    grade_class = f"grade-{grade.lower().replace('+', '-plus')}"
+    rec_badge = _get_recommendation_badge(grade, holding.recommended_action)
+    composite_score = holding.composite_score if holding.composite_score is not None else 0.0
+    rationale = holding.rationale_bullets[0] if holding.rationale_bullets else "Python analysis"
+    rationale_safe = escape(rationale)
+    badge = _confidence_badge(holding)
+    # Render fact pack provenance footer if the holding carries one (v5.2+)
+    fp = getattr(holding, "fact_pack", None)
+    fact_pack_footer = _fact_pack_provenance_footer(fp) if fp is not None else ""
+    # ADR-011: compact target/sell columns
+    target_cell = _format_target_cell(holding.price_targets, kind="target")
+    sell_cell = _format_target_cell(holding.price_targets, kind="sell")
+    return f"""
+        <tr>
+          <td><strong>{ticker_html}</strong><br><small>{name_safe}</small></td>
+          <td>{asset_class}</td>
+          <td class="{grade_class}"><strong>{grade}</strong>{badge}</td>
+          <td>{composite_score:.3f}</td>
+          <td>{rec_badge}</td>
+          <td><small>{rationale_safe}</small>{fact_pack_footer}</td>
+          {target_cell}
+          {sell_cell}
+        </tr>"""
+
+
+def generate_holdings_analysis(holdings: list[HoldingDecision]) -> str:
+    """Generate detailed holdings analysis."""
+    sorted_holdings = sorted(holdings, key=lambda h: (h.grade or "Z", -(h.composite_score or 0)))
+
+    holdings_rows = []
+    for holding in sorted_holdings:
+        holdings_rows.append(_render_holding_row(holding))
+
+    holdings_html = "".join(holdings_rows)
+
+    return f"""
+  <div class="section">
+    <h2>Detailed Holdings Analysis</h2>
+
+    <table>
+      <thead>
+        <tr>
+          <th>Ticker / Name</th>
+          <th>Class</th>
+          <th>Grade</th>
+          <th>Score</th>
+          <th>Recommendation</th>
+          <th>Rationale</th>
+          <th>Objectif (3-6 mo)</th>
+          <th>Niveau de vente</th>
+        </tr>
+      </thead>
+      <tbody>
+        {holdings_html}
+      </tbody>
+    </table>
+
+    <p class="small muted">Displaying all positions sorted by grade and score.</p>
+  </div>
+        """
+
+
+def generate_recommendations(portfolio_stats: dict[str, Any], discovery_results: dict[str, Any] | None = None) -> str:
+    """Generate recommendations section."""
+    a_plus_list = ""
+    if portfolio_stats.get("a_plus_holdings"):
+        a_plus_items = [
+            f"<strong>{getattr(h, 'ticker', 'N/A')}</strong> (Grade: {getattr(h, 'grade', 'N/A')}, Score: {getattr(h, 'composite_score', 0):.3f})"
+            for h in portfolio_stats["a_plus_holdings"]
+        ]
+        a_plus_list = f"""
+      <p><strong>A+ positions identified ({len(portfolio_stats["a_plus_holdings"])}):</strong></p>
+      <ul>
+        {"".join(f"<li>{item}</li>" for item in a_plus_items)}
+      </ul>"""
+
+    discovery_count = len(discovery_results["opportunities"]) if discovery_results and "opportunities" in discovery_results else 0
+
+    return f"""
+  <div class="section">
+    <h2>Strategic Recommendations</h2>
+
+    <div class="highlight warning">
+      <h3>Priority Actions</h3>
+      <ul>
+        <li><strong>Positions to sell:</strong> {portfolio_stats["recommendation_counts"]["SELL"]} positions need immediate attention</li>
+        <li><strong>A+ opportunities (current portfolio):</strong> {portfolio_stats["a_plus_count"]} excellent positions to keep or reinforce</li>
+        <li><strong>New opportunities discovered:</strong> {discovery_count} actionable candidates (grade C or better)</li>
+        <li><strong>Rebalancing:</strong> Consider diversification if excessive concentration</li>
+      </ul>
+      {a_plus_list}
+    </div>
+
+    <h3>Suggested Optimizations</h3>
+    <ul>
+      <li><strong>Risk reduction:</strong> Replace D/F positions with A/B alternatives</li>
+      <li><strong>Improve returns:</strong> Increase allocation to A+ positions</li>
+      <li><strong>Diversification:</strong> Balance between stocks, ETFs and crypto by risk profile</li>
+      <li><strong>Costs:</strong> Favor low-fee ETFs for passive exposure</li>
+    </ul>
+
+    <div class="highlight success">
+      <h3>Python Analysis Benefits</h3>
+      <ul>
+        <li><strong>Speed:</strong> Complete analysis in seconds vs minutes with AI</li>
+        <li><strong>Cost:</strong> $0 LLM fees vs $0.05-0.10 per analysis with AI</li>
+        <li><strong>Consistency:</strong> Identical results on every run</li>
+        <li><strong>Transparency:</strong> Auditable scoring algorithms</li>
+      </ul>
+    </div>
+  </div>
+        """
